Sync

+++ kernel/kernel/kernel.c
@@ -257,8 +257,8 @@ void kernel_main(multiboot_info_t *mbt)
     }
 
     nvmos_pagingOn(proc0Meta->pageDir);
-    printf("Test Mem 1: %p\n", *(uint32_t *)0x40000000);
-    printf("Test Mem 2: %p\n", *(uint32_t *)0x40001ffc);
+    printf("Test Mem 1: %p\n", (uint64_t) * (uint32_t *)0x40000000);
+    printf("Test Mem 2: %p\n", (uint64_t) * (uint32_t *)0x40001ffc);
 
 #undef allocTest
 endProc:
